Postpone CSV file write to program end
Records in a loaded CSV file previously did not need to be edited. Now
that a CSV file may contain a puzzle whose id is cached but had not yet
been solved, it's possible that a record may need to be edited. Thus,
rather than writing CSV records as we go, hold the database in-memory
and flush at the very end.

There's room to write out puzzles that _did_ have cached ids and solve
times, but writing everything at the end allows the data to be sorted,
which is nicer for human readability, even if less robust.

To simplify the database read/write, a user now only specifies one path
rather than a separate input and output path.